+++ src/lightning_habana/utils/imports.py
@@ -13,17 +13,18 @@
 # limitations under the License.
 
 import operator
-import os
 
-_PACKAGE_ROOT = os.path.dirname(__file__)
-_PROJECT_ROOT = os.path.dirname(_PACKAGE_ROOT)
+import torch
+from lightning_utilities.core.imports import RequirementCache, compare_version
+from packaging.version import Version
 
-from lightning_utilities.core.imports import compare_version, package_available  # noqa: E402
+from lightning_habana.utils.resources import _HABANA_FRAMEWORK_AVAILABLE, get_hpu_synapse_version  # noqa:
 
-_HABANA_FRAMEWORK_AVAILABLE = package_available("habana_frameworks")
-
+_HPU_SYNAPSE_GREATER_EQUAL_1_11_0 = Version(get_hpu_synapse_version()) >= Version("1.11.0")
 _TORCH_LESSER_EQUAL_1_13_1 = compare_version("torch", operator.le, "1.13.1")
 _TORCH_GREATER_EQUAL_2_0_0 = compare_version("torch", operator.ge, "2.0.0")
 _LIGHTNING_GREATER_EQUAL_2_0_0 = compare_version("lightning", operator.ge, "2.0.0") or compare_version(
     "pytorch_lightning", operator.ge, "2.0.0"
 )
+_TORCHVISION_AVAILABLE = RequirementCache("torchvision")
+_KINETO_AVAILABLE = torch.profiler.kineto_available()