+++ tests/test_pytorch/test_deepspeed.py
@@ -325,6 +325,48 @@ class SampleModel(LightningModule):
         return DataLoader(SampleDataset(16, 10), batch_size=2)
 
 
+def test_deepspeed_config(tmpdir):
+    """Test to ensure deepspeed config works correctly.
+
+    DeepSpeed config object including
+    optimizers/schedulers and saves the model weights to load correctly.
+    """
+
+    class TestCB(Callback):
+        def on_train_start(self, trainer, pl_module) -> None:
+            from torch.optim.lr_scheduler import StepLR
+
+            assert isinstance(trainer.optimizers[0], DeepSpeedZeroOptimizer)
+            assert isinstance(trainer.optimizers[0].optimizer, torch.optim.SGD)
+            assert isinstance(trainer.lr_scheduler_configs[0].scheduler, StepLR)
+            assert trainer.lr_scheduler_configs[0].interval == "epoch"
+
+    model = BoringModel()
+    lr_monitor = LearningRateMonitor()
+    _plugins = [DeepSpeedPrecisionPlugin(precision="bf16-mixed")]
+    trainer = Trainer(
+        accelerator=HPUAccelerator(),
+        strategy=HPUDeepSpeedStrategy(),
+        default_root_dir=tmpdir,
+        devices=1,
+        log_every_n_steps=1,
+        limit_train_batches=4,
+        limit_val_batches=4,
+        limit_test_batches=4,
+        max_epochs=2,
+        plugins=_plugins,
+        callbacks=[TestCB(), lr_monitor],
+        logger=CSVLogger(tmpdir),
+        enable_progress_bar=False,
+        enable_model_summary=False,
+    )
+
+    trainer.fit(model)
+    trainer.test(model)
+    assert list(lr_monitor.lrs) == ["lr-SGD"]
+    assert len(set(lr_monitor.lrs["lr-SGD"])) == trainer.max_epochs
+
+
 class SomeDataset(Dataset):
     def __init__(self, size, length):
         self.len = length
@@ -490,29 +532,55 @@ def test_lightning_deepspeed_stages(get_device_count, zero_stage, offload):
     trainer.fit(model)
 
 
-@pytest.mark.parametrize(
-    "remote_device", ["cpu", pytest.param("nvme", marks=pytest.mark.skip(reason="nvme is not supported currently"))]
-)
-def test_deepspeed_strategy_zero_infinity(tmpdir, get_device_count, remote_device):
-    model = SampleModel()
+def test_hpu_deepspeed_with_invalid_optimizer():
+    """Test to ensure if we pass an invalid optimizer and throws an exception."""
+
+    class DummyModel(BoringModel):
+        def configure_optimizers(self):
+            return None
+
+    import logging
+
+    model = DummyModel()
+    _plugins = [DeepSpeedPrecisionPlugin(precision="bf16-mixed")]
     trainer = Trainer(
         accelerator=HPUAccelerator(),
-        devices=get_device_count,
-        strategy=HPUDeepSpeedStrategy(
-            zero_optimization=True,
-            stage=3,
-            offload_optimizer=True,
-            offload_parameters=True,
-            remote_device=remote_device,
-            offload_params_device=remote_device,
-            offload_optimizer_device=remote_device,
-        ),
-        plugins=[DeepSpeedPrecisionPlugin(precision="bf16-mixed")],
-        fast_dev_run=8,
-        enable_progress_bar=False,
-        use_distributed_sampler=False,
-        limit_train_batches=16,
-        accumulate_grad_batches=2,
+        strategy=HPUDeepSpeedStrategy(logging_level=logging.INFO),
+        max_epochs=1,
+        plugins=_plugins,
+        devices=1,
     )
+    with pytest.raises(
+        MisconfigurationException, match="You have specified an invalid optimizer to be run with deepspeed."
+    ):
+        trainer.fit(model)
 
-    trainer.fit(model)
+
+def test_hpu_deepspeed_with_optimizer_and_config(deepspeed_zero_config):
+    """Test the preference of optimizer when configured both from deepspeed config and LightningModule."""
+
+    class DummyModel(BoringModel):
+        def configure_optimizers(self):
+            return torch.optim.AdamW(self.parameters(), lr=0.1)
+
+    class TestCB(Callback):
+        def on_train_start(self, trainer, pl_module) -> None:
+            from deepspeed.runtime.lr_schedules import WarmupLR
+
+            assert isinstance(trainer.optimizers[0], DeepSpeedZeroOptimizer)
+            assert isinstance(trainer.optimizers[0].optimizer, torch.optim.AdamW)
+            assert isinstance(trainer.lr_scheduler_configs[0].scheduler, WarmupLR)
+            assert trainer.lr_scheduler_configs[0].interval == "step"
+
+    import logging
+
+    model = DummyModel()
+    _plugins = [DeepSpeedPrecisionPlugin(precision="bf16-mixed")]
+    trainer = Trainer(
+        accelerator=HPUAccelerator(),
+        strategy=HPUDeepSpeedStrategy(logging_level=logging.INFO, config=deepspeed_zero_config),
+        callbacks=[TestCB()],
+        max_epochs=1,
+        plugins=_plugins,
+        devices=1,
+    )